extract Amenity menu building to helper. refacoring.
